rename files to fix typo
